Split battle_init into 3 files
